+++ Commands/Admin/NewTenantSite.cs
@@ -1,25 +1,34 @@
-﻿using System.Management.Automation;
+﻿using System;
+using System.Management.Automation;
 using Microsoft.SharePoint.Client;
+using OfficeDevPnP.Core;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using SharePointPnP.PowerShell.Commands.Base;
 using OfficeDevPnP.Core.Entities;
+using Resources = SharePointPnP.PowerShell.Commands.Properties.Resources;
+
 
 namespace SharePointPnP.PowerShell.Commands
 {
     [Cmdlet(VerbsCommon.New, "PnPTenantSite")]
     [CmdletAlias("New-SPOTenantSite")]
-    [CmdletHelp("Creates a new site collection for the current tenant", 
+    [CmdletHelp("Creates a new site collection for the current tenant",
         DetailedDescription = @"The New-PnPTenantSite cmdlet creates a new site collection for the current company. However, creating a new SharePoint
-Online site collection fails if a deleted site with the same URL exists in the Recycle Bin. If you want to use this command for an on-premises farm, please refer to http://blogs.msdn.com/b/vesku/archive/2014/06/09/provisioning-site-collections-using-sp-app-model-in-on-premises-with-just-csom.aspx ", 
+Online site collection fails if a deleted site with the same URL exists in the Recycle Bin. If you want to use this command for an on-premises farm, please refer to http://blogs.msdn.com/b/vesku/archive/2014/06/09/provisioning-site-collections-using-sp-app-model-in-on-premises-with-just-csom.aspx ",
         Category = CmdletHelpCategory.TenantAdmin)]
     [CmdletExample(
-        Code = @"PS:> New-PnPTenantSite -Title Contoso -Url https://tenant.sharepoint.com/sites/contoso -Owner user@example.org -TimeZone 4",
-        Remarks = @"This will add a site collection with the title 'Contoso', the url 'https://tenant.sharepoint.com/sites/contoso', the timezone 'UTC+01:00' and the owner 'user@example.org'", SortOrder = 1)]        
+        Code = @"PS:> New-PnPTenantSite -Title Contoso -Url https://tenant.sharepoint.com/sites/contoso -Owner user@example.org -TimeZone 4 -Template STS#0",
+        Remarks = @"This will add a site collection with the title 'Contoso', the url 'https://tenant.sharepoint.com/sites/contoso', the timezone 'UTC+01:00',the owner 'user@example.org' and the template used will be STS#0, a TeamSite", 
+        SortOrder = 1)]
+    [CmdletExample(
+        Code = @"PS:> New-PnPTenantSite -Title Contoso -Url /sites/contososite -Owner user@example.org -TimeZone 4 -Template STS#0",
+        Remarks = @"This will add a site collection with the title 'Contoso', the url 'https://tenant.sharepoint.com/sites/contososite' of which the base part will be picked up from your current connection, the timezone 'UTC+01:00', the owner 'user@example.org' and the template used will be STS#0, a TeamSite", 
+        SortOrder = 2)]
     [CmdletRelatedLink(
-        Text ="Locale IDs", 
+        Text = "Locale IDs",
         Url = "http://go.microsoft.com/fwlink/p/?LinkId=242911Id=242911")]
     [CmdletRelatedLink(
-        Text = "Resource Usage Limits on Sandboxed Solutions in SharePoint 2010", 
+        Text = "Resource Usage Limits on Sandboxed Solutions in SharePoint 2010",
         Url = "http://msdn.microsoft.com/en-us/library/gg615462.aspx.")]
     [CmdletRelatedLink(
         Text = "Creating on-premises site collections using CSOM",
@@ -66,26 +75,54 @@ Online site collection fails if a deleted site with the same URL exists in the R
         [Parameter(Mandatory = false)]
         public SwitchParameter Wait;
 
+        [Parameter(Mandatory = false, HelpMessage = "Do not ask for confirmation.")]
+        public SwitchParameter Force;
+
         protected override void ExecuteCmdlet()
         {
+            bool shouldContinue = true;
+            if (!Url.ToLower().StartsWith("https://") && !Url.ToLower().StartsWith("http://"))
+            {
+                Uri uri = BaseUri;
+                Url = $"{uri.ToString().TrimEnd('/')}/{Url.TrimStart('/')}";
+                shouldContinue = ShouldContinue(string.Format(Resources.CreateSiteWithUrl0, Url), Resources.Confirm);
+            }
+            if (Force || shouldContinue)
+            {
 #if ONPREMISES
-            var entity = new SiteEntity();
-            entity.Url = Url;
-            entity.Title = Title;
-            entity.SiteOwnerLogin = Owner;
-            entity.Template = Template;
-            entity.StorageMaximumLevel = StorageQuota;
-            entity.StorageWarningLevel = StorageQuotaWarningLevel;
-            entity.TimeZoneId = TimeZone;
-            entity.UserCodeMaximumLevel = ResourceQuota;
-            entity.UserCodeWarningLevel = ResourceQuotaWarningLevel;
-            entity.Lcid = Lcid;
+                var entity = new SiteEntity();
+                entity.Url = Url;
+                entity.Title = Title;
+                entity.SiteOwnerLogin = Owner;
+                entity.Template = Template;
+                entity.StorageMaximumLevel = StorageQuota;
+                entity.StorageWarningLevel = StorageQuotaWarningLevel;
+                entity.TimeZoneId = TimeZone;
+                entity.UserCodeMaximumLevel = ResourceQuota;
+                entity.UserCodeWarningLevel = ResourceQuotaWarningLevel;
+                entity.Lcid = Lcid;
 
-            Tenant.CreateSiteCollection(entity);
+                Tenant.CreateSiteCollection(entity);
 #else
-            Tenant.CreateSiteCollection(Url, Title, Owner, Template, (int)StorageQuota, (int)StorageQuotaWarningLevel, TimeZone, (int)ResourceQuota, (int)ResourceQuotaWarningLevel, Lcid, RemoveDeletedSite, Wait);
+                Func<TenantOperationMessage, bool> timeoutFunction = TimeoutFunction;
+
+
+                Tenant.CreateSiteCollection(Url, Title, Owner, Template, (int)StorageQuota,
+                    (int)StorageQuotaWarningLevel, TimeZone, (int)ResourceQuota, (int)ResourceQuotaWarningLevel, Lcid,
+                    RemoveDeletedSite, Wait, Wait == true ? timeoutFunction : null);
 #endif
+            }
         }
 
+#if !ONPREMISES
+        private bool TimeoutFunction(TenantOperationMessage message)
+        {
+            if (message == TenantOperationMessage.CreatingSiteCollection)
+            {
+                Host.UI.Write(".");
+            }
+            return Stopping;
+        }
+#endif
     }
 }