+++ Commands/Files/SetFileCheckedOut.cs
@@ -9,9 +9,9 @@ namespace SharePointPnP.PowerShell.Commands.Files
     [CmdletHelp("Checks out a file",
         Category = CmdletHelpCategory.Files)]
     [CmdletExample(
-        Code = @"PS:>Set-PnPFileCheckedOut -Url ""/Documents/Contract.docx""",
+        Code = @"PS:>Set-PnPFileCheckedOut -Url ""/sites/testsite/subsite/Documents/Contract.docx""",
         SortOrder = 1,
-        Remarks = @"Checks out the file ""Contract.docx"" in the ""Documents"" library")]
+        Remarks = @"Checks out the file ""Contract.docx"" in the ""Documents"" library.")]
     public class SetFileCheckedOut : SPOWebCmdlet
     {
         [Parameter(Mandatory = true, Position=0, ValueFromPipeline=true, HelpMessage = @"The server relative url of the file to check out")]