+++ Commands/Lists/SetList.cs
@@ -83,7 +83,7 @@ namespace SharePointPnP.PowerShell.Commands.Lists
                     isDirty = true;
                 }
 
-                if (list.ContentTypesEnabled != EnableContentTypes)
+                if (MyInvocation.BoundParameters.ContainsKey("EnableContentTypes") && list.ContentTypesEnabled != EnableContentTypes)
                 {
                     list.ContentTypesEnabled = EnableContentTypes;
                     isDirty = true;