+++ Commands/Properties/Resources.Designer.cs
@@ -70,6 +70,42 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Permanently delete all items in the first and second stage recycle bins?.
+        /// </summary>
+        internal static string ClearBothRecycleBins {
+            get {
+                return ResourceManager.GetString("ClearBothRecycleBins", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Permanently delete file &apos;{0}&apos; from the recycle bin?.
+        /// </summary>
+        internal static string ClearRecycleBinItem {
+            get {
+                return ResourceManager.GetString("ClearRecycleBinItem", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Permanently delete all the items in the second stage recycle bin?.
+        /// </summary>
+        internal static string ClearSecondStageRecycleBin {
+            get {
+                return ResourceManager.GetString("ClearSecondStageRecycleBin", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Permanently delete the site collection previously located at &apos;{0}&apos; from the recycle bin?.
+        /// </summary>
+        internal static string ClearTenantRecycleBinItem {
+            get {
+                return ResourceManager.GetString("ClearTenantRecycleBinItem", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Command not executed due to healthscore value of {0}.
         /// </summary>
         internal static string CommandNotExecutedDueToHealthscoreValueOf0 {
@@ -93,6 +129,15 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         internal static string CopyFile0To1 {
             get {
                 return ResourceManager.GetString("CopyFile0To1", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Create site with url &apos;{0}&apos;?.
+        /// </summary>
+        internal static string CreateSiteWithUrl0 {
+            get {
+                return ResourceManager.GetString("CreateSiteWithUrl0", resourceCulture);
             }
         }
         
@@ -206,6 +251,24 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Move all items from the first stage recycle bin to the second stage recycle bin?.
+        /// </summary>
+        internal static string MoveFirstStageRecycleBinItemsToSecondStage {
+            get {
+                return ResourceManager.GetString("MoveFirstStageRecycleBinItemsToSecondStage", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Move recycle bin item with leaf &apos;{0}&apos; to the second stage recycle bin?.
+        /// </summary>
+        internal static string MoveRecycleBinItemWithLeaf0ToSecondStage {
+            get {
+                return ResourceManager.GetString("MoveRecycleBinItemWithLeaf0ToSecondStage", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to No Azure AD connection, please connect first with Connect-PnPMicrosoftGraph.
         /// </summary>
         internal static string NoAzureADAccessToken {
@@ -293,6 +356,15 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         internal static string RemoveGroup0 {
             get {
                 return ResourceManager.GetString("RemoveGroup0", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Remove Image Rendition with name &apos;{0}&apos;?.
+        /// </summary>
+        internal static string RemoveImageRenditionWithName0 {
+            get {
+                return ResourceManager.GetString("RemoveImageRenditionWithName0", resourceCulture);
             }
         }
         
@@ -419,6 +491,33 @@ namespace SharePointPnP.PowerShell.Commands.Properties {
         internal static string RenameFile0To1 {
             get {
                 return ResourceManager.GetString("RenameFile0To1", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Restore the site collection previously located at &apos;{0}&apos; from the recycle bin to its original location?.
+        /// </summary>
+        internal static string ResetTenantRecycleBinItem {
+            get {
+                return ResourceManager.GetString("ResetTenantRecycleBinItem", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Restore the file &apos;{0}&apos; from the recycle bin to its original location?.
+        /// </summary>
+        internal static string RestoreRecycleBinItem {
+            get {
+                return ResourceManager.GetString("RestoreRecycleBinItem", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Restore all items from the first and second stage recycle bins to their original locations?.
+        /// </summary>
+        internal static string RestoreRecycleBinItems {
+            get {
+                return ResourceManager.GetString("RestoreRecycleBinItems", resourceCulture);
             }
         }