+++ changelog.md
@@ -1,5 +1,17 @@
 # SharePointPnP.PowerShell Changelog #
 
+**2017-01-05**
+* Added Get-PnPTenantRecyclyBinItem cmdlet to list all items in the tenant scoped recycle bin
+* Added -Wait and -LockState properties to Set-PnPTenantSite
+* The Tenant cmdlets now report progress if the -Wait parameter is specified (where applicable)
+
+**2017-01-03**
+* HAPPY NEW YEAR!
+* Added Clear-PnPRecyclyBinItem, Clear-PnPTenantRecyclyBinItem, Get-PnPRecyclyBinItem, Move-PnPRecyclyBinItem, Restore-PnPRecyclyBinItem, Restore-PnPTenantRecyclyBinItem cmdlets
+* Added Move-PnPFolder, Rename-PnPFolder cmdlets
+* Added Add-PnPPublishingImageRendition, Get-PnPPublishingImageRendition and Remove-PnPPublishingImageRendition cmdlets
+* Refactored Get-PnPFile. ServerRelativeUrl and SiteRelativeUrl are now obsolete (but will still work), use the Url parameter instead which takes either a server or site relative url.
+
 **2016-11-21**
 * Added support to enable versionining and set the maximum number of versions to keep on a list and library with Set-PnPList
 * Updated Add-PnPUserToGroup to allow to send invites to external users